feat: Use integers for histogram bins and add fingerprint fields

+++ src/main/kotlin/org/wfanet/measurement/reporting/job/BasicReportsReportsJob.kt
@@ -666,7 +666,7 @@ class BasicReportsReportsJob(
       .NoisyMetricSetKt
       .histogramResult {
         for (bin: MetricResult.HistogramResult.Bin in source.binsList) {
-          binResults[bin.label] =
+          binResults[bin.label.toInt()] =
             ReportResultKt.ReportingSetResultKt.ReportingWindowResultKt.NoisyReportResultValuesKt
               .NoisyMetricSetKt
               .HistogramResultKt
